@@ -7,7 +7,7 @@ from torchvision.models.vgg import VGG16_Weights
 
 class Vgg16(torch.nn.Module):
     def __init__(self, requires_grad=False):
-        super(Vgg16, self).__init__()
+        super().__init__()
         vgg_pretrained_features = models.vgg16(weights=VGG16_Weights.IMAGENET1K_V1).features
         self.slice1 = torch.nn.Sequential()
         self.slice2 = torch.nn.Sequential()
