@@ -64,7 +64,7 @@ class Generator(Net):
     """
 
     def __init__(self, z_dim, nf, nc):
-        super(Generator, self).__init__()
+        super().__init__()
 
         self.net = nn.Sequential(
             # input is Z, going into a convolution
@@ -103,7 +103,7 @@ class Discriminator(Net):
     """
 
     def __init__(self, nc, nf):
-        super(Discriminator, self).__init__()
+        super().__init__()
 
         self.net = nn.Sequential(
             # input is (nc) x 64 x 64
