@@ -21,7 +21,7 @@ eps = np.finfo(np.float32).eps.item()
 
 class Policy(nn.Module):
     def __init__(self):
-        super(Policy, self).__init__()
+        super().__init__()
         self.affine1 = nn.Linear(4, 128)
         self.dropout = nn.Dropout(p=0.6)
         self.affine2 = nn.Linear(128, 2)
