@@ -272,7 +272,7 @@ class _OutputHandler(BaseOutputHandler):
         if metric_names is None and output_transform is None:
             # This helps to avoid 'Either metric_names or output_transform should be defined' of BaseOutputHandler
             metric_names = []
-        super(_OutputHandler, self).__init__(
+        super().__init__(
             description, metric_names, output_transform, global_step_transform=None, state_attributes=state_attributes
         )
         self.closing_event_name = closing_event_name
