@@ -13,7 +13,7 @@ from ignite.metrics import Accuracy, Metric, MetricsLambda, Precision, Recall
 
 class ListGatherMetric(Metric):
     def __init__(self, index):
-        super(ListGatherMetric, self).__init__()
+        super().__init__()
         self.index = index
 
     def reset(self):
